refactor(tests): consolidate duplicated fixtures and helpers (E1-E3)
E1: Single mock_db_api fixture in tests/helpers/fixtures.py, imported
via root conftest. Replaces 3 drifting copies (3/4/5 stubs) with one
canonical definition (5 stubs). core_proxy also moved to helpers.

E2: Extract is_logout_request to sf_core/tests/common/mocks/session.rs.
Removes identical 5-line function from e2e and integration logout tests.

E3: Log catch_unwind panic payload in SnowflakeTestClient::drop instead
of silently discarding it with let _ =. Uses eprintln! (not tracing)
since the subscriber may be torn down during the panic.
