+++ sf_core/tests/common/mocks/session.rs
@@ -1,0 +1,8 @@
+/// Check whether a wiremock request is a logout (DELETE session) request.
+pub fn is_logout_request(r: &wiremock::Request) -> bool {
+    r.url.path() == "/session"
+        && r.url
+            .query()
+            .map(|q| q.contains("delete=true"))
+            .unwrap_or(false)
+}
